# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /howto/{id}:
    get:
      tags: [How-to]
      operationId: getPublishedHowto
      summary: Get a published How-to entry
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Published entry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HowtoEntry' }
        '404': { description: Entry not found or not published. }
  /admin/howto:
    get:
      tags: [How-to]
      operationId: listAdminHowto
      summary: List all How-to entries for editors
      responses:
        '200':
          description: Editorial entries.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/HowtoEntry' }
    post:
      tags: [How-to]
      operationId: createAdminHowto
      summary: Create a How-to entry
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HowtoEntryInput' }
      responses:
        '201':
          description: Entry created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HowtoEntry' }
        '400': { description: Invalid entry. }
        '409': { description: Entry id already exists. }
  /admin/howto/{id}:
    get:
      tags: [How-to]
      operationId: getAdminHowto
      summary: Get an entry for editing
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Editorial entry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HowtoEntry' }
        '404': { description: Entry not found. }
    patch:
      tags: [How-to]
      operationId: updateAdminHowto
      summary: Partially update a How-to entry
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HowtoEntryPatch' }
      responses:
        '200':
          description: Updated entry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HowtoEntry' }
        '400': { description: Invalid entry. }
        '404': { description: Entry not found. }
    delete:
      tags: [How-to]
      operationId: deleteAdminHowto
      summary: Delete a How-to entry
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Entry deleted. }
        '404': { description: Entry not found. }
...
components:
...
  schemas:
...
    HowtoInlineRun:
      type: object
      additionalProperties: false
      required: [text]
      properties:
        text: { type: string, maxLength: 12000 }
        bold: { type: boolean }
        italic: { type: boolean }
        href: { type: string, format: uri }
    HowtoBlock:
      oneOf:
        - type: object
          additionalProperties: false
          required: [type, runs]
          properties:
            type: { type: string, enum: [heading, paragraph, tip] }
            runs: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/HowtoInlineRun' } }
        - type: object
          additionalProperties: false
          required: [type, number, title, runs]
          properties:
            type: { type: string, const: step }
            number: { type: integer, minimum: 1, maximum: 100 }
            title: { type: string, maxLength: 300 }
            runs: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/HowtoInlineRun' } }
        - type: object
          additionalProperties: false
          required: [type, items]
          properties:
            type: { type: string, const: list }
            items: { type: array, maxItems: 100, items: { type: string, maxLength: 4000 } }
        - type: object
          additionalProperties: false
          required: [type, mediaId, altText]
          properties:
            type: { type: string, const: image }
            mediaId: { type: string, minLength: 1, maxLength: 200 }
            altText: { type: string, maxLength: 1000 }
            caption: { type: string, maxLength: 2000 }
            url: { type: [string, 'null'] }
        - type: object
          additionalProperties: false
          required: [type, url]
          properties:
            type: { type: string, const: video }
            url: { type: string, format: uri, maxLength: 2000 }
            caption: { type: string, maxLength: 2000 }
    HowtoEntry:
      type: object
      required: [id, title, description, type, readTime, displayOrder, status, body, createdAt, updatedAt, publishedAt]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        type: { type: string, enum: [Article, Guide, Video] }
        readTime: { type: string }
        displayOrder: { type: integer }
        status: { type: string, enum: [draft, published] }
        body: { type: array, maxItems: 300, items: { $ref: '#/components/schemas/HowtoBlock' } }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        publishedAt: { type: [string, 'null'], format: date-time }
    HowtoEntryInput:
      type: object
      additionalProperties: false
      required: [id, title, description, type, readTime, displayOrder, status, body]
      properties:
        id: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', maxLength: 100 }
        title: { type: string, minLength: 1, maxLength: 300 }
        description: { type: string, maxLength: 2000 }
        type: { type: string, enum: [Article, Guide, Video] }
        readTime: { type: string, maxLength: 100 }
        displayOrder: { type: integer, minimum: 0, maximum: 1000000 }
        status: { type: string, enum: [draft, published] }
        body: { type: array, maxItems: 300, items: { $ref: '#/components/schemas/HowtoBlock' } }
    HowtoEntryPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title: { type: string, minLength: 1, maxLength: 300 }
        description: { type: string, maxLength: 2000 }
        type: { type: string, enum: [Article, Guide, Video] }
        readTime: { type: string, maxLength: 100 }
        displayOrder: { type: integer, minimum: 0, maximum: 1000000 }
        status: { type: string, enum: [draft, published] }
        body: { type: array, maxItems: 300, items: { $ref: '#/components/schemas/HowtoBlock' } }
...