# lib/api-zod/src/generated/types/howtoBlock.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HowtoInlineRun } from './howtoInlineRun';

export type HowtoBlock = {
  type: 'heading' | 'paragraph' | 'tip';
  /** @maxItems 100 */
  runs: HowtoInlineRun[];
} | {
  type: 'step';
  /**
     * @minimum 1
     * @maximum 100
     */
  number: number;
  /** @maxLength 300 */
  title: string;
  /** @maxItems 100 */
  runs: HowtoInlineRun[];
} | {
  type: 'list';
  /**
     * @maxItems 100
     * @items.maxLength 4000
     */
  items: string[];
} | {
  type: 'image';
  /**
     * @minLength 1
     * @maxLength 200
     */
  mediaId: string;
  /** @maxLength 1000 */
  altText: string;
  /** @maxLength 2000 */
  caption?: string;
  /** @nullable */
  url?: string | null;
} | {
  type: 'video';
  /** @maxLength 2000 */
  url: string;
  /** @maxLength 2000 */
  caption?: string;
};