# lib/api-zod/src/generated/types/howtoEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HowtoBlock } from './howtoBlock';
import type { HowtoEntryStatus } from './howtoEntryStatus';
import type { HowtoEntryType } from './howtoEntryType';

export interface HowtoEntry {
  id: string;
  title: string;
  description: string;
  type: HowtoEntryType;
  readTime: string;
  displayOrder: number;
  status: HowtoEntryStatus;
  /** @maxItems 300 */
  body: HowtoBlock[];
  createdAt: Date;
  updatedAt: Date;
  /** @nullable */
  publishedAt: Date | null;
}