# lib/api-zod/src/generated/types/howtoEntryInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HowtoBlock } from './howtoBlock';
import type { HowtoEntryInputStatus } from './howtoEntryInputStatus';
import type { HowtoEntryInputType } from './howtoEntryInputType';

export interface HowtoEntryInput {
  /**
     * @maxLength 100
     * @pattern ^[a-z0-9]+(?:-[a-z0-9]+)*$
     */
  id: string;
  /**
     * @minLength 1
     * @maxLength 300
     */
  title: string;
  /** @maxLength 2000 */
  description: string;
  type: HowtoEntryInputType;
  /** @maxLength 100 */
  readTime: string;
  /**
     * @minimum 0
     * @maximum 1000000
     */
  displayOrder: number;
  status: HowtoEntryInputStatus;
  /** @maxItems 300 */
  body: HowtoBlock[];
}