# lib/api-zod/src/generated/types/howtoEntryPatch.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HowtoBlock } from './howtoBlock';
import type { HowtoEntryPatchStatus } from './howtoEntryPatchStatus';
import type { HowtoEntryPatchType } from './howtoEntryPatchType';

export interface HowtoEntryPatch {
  /**
     * @minLength 1
     * @maxLength 300
     */
  title?: string;
  /** @maxLength 2000 */
  description?: string;
  type?: HowtoEntryPatchType;
  /** @maxLength 100 */
  readTime?: string;
  /**
     * @minimum 0
     * @maximum 1000000
     */
  displayOrder?: number;
  status?: HowtoEntryPatchStatus;
  /** @maxItems 300 */
  body?: HowtoBlock[];
}