# lib/api-zod/src/generated/types/howtoInlineRun.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HowtoInlineRun {
  /** @maxLength 12000 */
  text: string;
  bold?: boolean;
  italic?: boolean;
  href?: string;
}